Properly implement deletion of local words
Deletion of a local word can mean the deletion of the whole record if
dealing with both grades. Or it cxan mean to set either the contracted
or uncontracted column to NULL if just dealing with one grade. And of
course deleting the whole record if a partial delete resulted in both
columns beeing NULL.