Remove nil values from local words
If a word only has braille for contracted defined for example then the
:uncontracted key is not defined for that word (as opposed to have the
key with a nil value).

However as soon as you store the value in the db then the column
uncontracted will be NULL.

So if you fetch the word from the db the returned value will have the
:uncontracted with a nil value. This is valid but not what we want as
the rest of the code assumes a key to be absent if the value is not
defined.

Therefore we remove keys where the value is nil in local/get-words.