Move from optional keys to nilable keys
In other words we use the same concept as the database. It a
value (contracted or uncontracted) is not given then the key is still
there but the value will be null.

This simplifies quite a bit in the interaction with the database. The
only quirk is that when sending delete requests from the browser we
need to dissoc some parts of the word depending on the given grade.