Give saved words a new key so that component localk state is refreshed
The input fields of the words have component local state such as
initial value. This is used for a simple undo mechanism. Now this
local state should be invalidated when the word is actually saved (as
we have a new initial value now). Since this component local state is
implemented using a closure it cannot be changed fom outside. We would
either have to implement the local state globally so that it could be
changed from a save or the component would have to be recreated with
the new initial value. This is described in #81.

We choose option two by just assigning a new key to the word which
essentially forces react to create a new component which will have a
now closure with the right initial values.

Fixes #81