Remove subscriptions from event handlers
Fixes #64

Subscriptions were used as a sort of an accessor function in event
handlers. This doesn't make sense. So the affected subscriptions are
now spit into an accessor function and the actual subscription
itself (which uses the accessor). The event handlers now just use the
accessor function which grabs the right data out of the app state db.