Move to a much simpler model for the online player
Simply open the link to the online player in a separate window. That
way we no longer need a spinner on the button and we no longer need to
know when the epub is ready (to disable the spinner). We can simply
return a 302 from the original request (that created the epub) and the
browser is transparently opening the online player url that we are
redirecting to.