Implement unignore
at least a very simple version of it.

To be truly useful we'd probably have to do a re-fetch of the words
after an unignore to make sure an unignored words does not just
disapear from the view (until the next time the words are re-fetched)

Maybe the concept of unignoring doesn't make any sense. We should
probably just clear the isIgnored field once a word has been
confirmed.