Move the materialization of the state column to the db layer
i.e. the queries in `queries.sql` will do the necessary joins so that
all returned documents will already contain a state column. Also the
update of documents query will handle state names and convert them to
the appropriate state_id.

This is a major simplification of the UI where we no longer have to
know that there is a dichtonomy between states and state-ids. We can
now simply deal with states and do not have to do any conversion.

Also the brittle mapping between state and state-id is now
concentrated at the db layer.