Handle preview differently: Return a 201 (created) with the location
of the created new resource. The client can then fetch the generated
resource in a second request and possibly open it in a different
window.

There is a new route to serve generated assets. These are stored in a
spool directory and are statically served from there. There should
probably be some kind of expiry mechanism.

When requesting a generated resource the api just returns a 201 with
the location of the generated resource. That way we can keep the api
pure JSON and do the download via a different channel.