Use Failjure to indicate validation errors instead of exceptions
While Exceptions are a nice way to get early return in Clojure they
indeed mix control flow and error handling. This was getting apparent
when instrumenting the function insert-version. Since validation
problems were signaled using exceptions the alerting of prometheus was
also triggered.

To avoid this we now use Adam Bards Failjure to return a failure
object in case of validation problems. The return value is then
checked for a failure and handled accordingly without resorting to
Exceptions.