Make sure mixed emphasis words are kept together
or generally do not split words at inline element boundaries

e.g. CO<sup>2</sup> will become CO2

This is useless in the dictionary as liblouis translates each node
separately, i.e. translates 'CO' and '2' separately. So there is no
real point in adding CO2 to the dictionary.

OTOH there is also no point in putting the separate words 'CO' and '2'
in the dictionary, as the actual string that is sent to liblouis in
this case of mixed emphasis contains additional characters. So putting
the separate words in the dictionary has also no effect.

The best would probably be to not show mixed emphasis words in the
dictionary altogether.