Handle failure of braille translation of unknown words
When the braille translation of an unknown word fails we have a
problem.

We simply do not add any braille to that particular word. The UI will
then disable all controlls on that word, i.e. you will not be able to
save it and.

To get rid of the word you will have to wrap it in a brl:select and
re-upload the xmlp