Make latex conversion with images work
Previously the latex process was invoked in a tempdir where the images
were copied to. However the latex file itself was kept is a separate
temporary location. So, of course, the image paths were not correct.

Now the latex file is copied into the same tempdir where the images
are copied to.

Also use the `with-tempdir` macro to ensure deletion of temporary
files.